Conflicts:
	backend/src/main/nl/tudelft/ti2806/riverrush/controller/RenderController.java
	backend/src/main/nl/tudelft/ti2806/riverrush/controller/UserController.java
	backend/src/main/nl/tudelft/ti2806/riverrush/domain/entity/state/AnimalInAir.java
	backend/src/main/nl/tudelft/ti2806/riverrush/game/Game.java
	backend/src/main/nl/tudelft/ti2806/riverrush/game/GameTrack.java
	backend/src/main/nl/tudelft/ti2806/riverrush/game/state/PlayingGameState.java
	backend/src/test/nl/tudelft/ti2806/riverrush/game/GameTrackTest.java
	core/src/main/nl/tudelft/ti2806/riverrush/domain/entity/AbstractAnimal.java
	core/src/main/nl/tudelft/ti2806/riverrush/domain/event/BasicEventDispatcher.java
	desktop/src/main/nl/tudelft/ti2806/riverrush/controller/RenderController.java
	desktop/src/main/nl/tudelft/ti2806/riverrush/desktop/MainDesktop.java
	desktop/src/main/nl/tudelft/ti2806/riverrush/game/state/PlayingGameState.java
	desktop/src/main/nl/tudelft/ti2806/riverrush/graphics/entity/AnimalActor.java
	desktop/src/main/nl/tudelft/ti2806/riverrush/graphics/entity/BoatGroup.java
	desktop/src/main/nl/tudelft/ti2806/riverrush/graphics/entity/BoatSector.java
	desktop/src/main/nl/tudelft/ti2806/riverrush/graphics/entity/state/AnimalOnBoat.java
	desktop/src/main/nl/tudelft/ti2806/riverrush/screen/PlayingGameScreen.java

Made adding animals work again
Renamed some stuff
Cried a bit
Lost the will to live

+++ core/src/main/nl/tudelft/ti2806/riverrush/domain/event/TeamProgressEvent.java
@@ -21,7 +21,7 @@ public class TeamProgressEvent implements Event {
     @Override
     public Event deserialize(final Map<String, String> keyValuePairs) {
         teamID = Integer.parseInt(keyValuePairs.get("teamID"));
-        progress = Integer.parseInt(keyValuePairs.get("progress"));
+        progress = Double.parseDouble(keyValuePairs.get("progress"));
         return this;
     }
